requirements.txt
a2wsgi==1.10.8
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
aiosqlite==0.21.0
alabaster==1.0.0
annotated-types==0.7.0
anyio==4.8.0
appdirs==1.4.4
asgiref==3.8.1
attrs==25.1.0
babel==2.17.0
bentoml==1.3.21
cattrs==23.1.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-option-group==0.5.6
cloudpickle==3.1.1
coverage==7.6.11
deepmerge==2.0
Deprecated==1.2.18
docutils==0.21.2
flake8==7.1.1
frozenlist==1.5.0
fs==2.4.16
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-ws==0.7.1
idna==3.10
imagesize==1.4.1
imbalanced-learn==0.13.0
importlib_metadata==8.5.0
inflection==0.5.1
iniconfig==2.0.0
Jinja2==3.1.5
joblib==1.4.2
kantoku==0.18.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
multidict==6.1.0
numpy==2.2.2
nvidia-ml-py==12.570.86
opentelemetry-api==1.30.0
opentelemetry-instrumentation==0.51b0
opentelemetry-instrumentation-aiohttp-client==0.51b0
opentelemetry-instrumentation-asgi==0.51b0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
opentelemetry-util-http==0.51b0
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pathspec==0.12.1
pip-requirements-parser==32.0.1
pluggy==1.5.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.2.1
psutil==6.1.1
pycodestyle==2.12.1
pydantic==2.10.6
pydantic_core==2.27.2
pyflakes==3.2.0
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.2.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
questionary==2.1.0
requests==2.32.3
rich==13.9.4
schema==0.7.7
scikit-learn==1.6.1
scipy==1.15.1
setuptools==75.8.0
simple-di==0.1.5
six==1.17.0
sklearn-compat==0.1.3
sniffio==1.3.1
snowballstemmer==2.2.0
Sphinx==8.1.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
-e git+https://github.com/DataScientest-Studio/Template_MLOps_accidents.git@07ee5005fe2f0607847c3057acffdc04785aac25#egg=src
starlette==0.45.3
threadpoolctl==3.5.0
tomli_w==1.2.0
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uv==0.5.29
uvicorn==0.34.0
watchfiles==1.0.4
wcwidth==0.2.13
wrapt==1.17.2
wsproto==1.2.0
yarl==1.18.3
zipp==3.21.0
